Refactor CommunityService and fix typos
Refactored the code in CommunityService to use async/await and enhance readability. Added a new method, getPostsFromCommunity, to retrieve posts specific to a community. Typos were fixed in the CommentController. A GarDbContext was added in Program.cs. Also, modifications were made to improve naming conventions in GetCommunityById in the CommunityRepository.