Refactor service methods to be asynchronous
The modification primarily affects several service methods to run as asynchronous tasks for increased efficiency in performance. All the tasks previously returning Task<TResult> are now updated to use the async keyword returning TResult, which makes the methods awaitable. Moreover, methods accessing the database directly are refactored to utilize new extension methods for better organization of the code and an increase in modularity. Some additional modifications also include improved mapping configurations and a slight optimization in Login and Register endpoints.